% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInit.R
\name{checkInit}
\alias{checkInit}
\title{Checks log book has been initialised.}
\usage{
checkInit()
}
\value{
None.  An error if initialisation not complete.
}
\description{
Checks that the required initialisation parameters exist and are sane.
}
