% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initLogBook.R
\name{initLogBook}
\alias{initLogBook}
\title{Initialise the logBook session.}
\usage{
initLogBook(baseDir, ...)
}
\arguments{
\item{baseDir}{The top level directory where sessions are logged. Can be omitted if set by RLOGBOOK_baseDir or an earlier call to \code{initLogBook}.}

\item{...}{Extra parameters to be set.}
}
\value{
Nothing, but a new session is created and plots can now be logged.
}
\description{
Initialises log Book parameters, starts a "session" (which is basically just a directory to store stuff).  This means making a directory.  Parameters are set in the following order:
 - package defaults.
 - previously set values in this R session.
 - values from RLOGBOOK_<> environmental variables.
 - values passed to this function.
}
