% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{initParams}
\alias{initParams}
\title{Initialise parameters.  Internal function.}
\usage{
initParams(toSet)
}
\arguments{
\item{toSet}{List of parameters to be set.}
}
\value{
The params object is updated, but nothing is returned.
}
\description{
Initialise parameters used by logBook.  See \code{\link{initLogBook}} for order of precedence when setting params.
}
