% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPlot.R
\name{op}
\alias{op}
\title{Open a new plot}
\usage{
op(plotName, ...)
}
\arguments{
\item{plotName}{Extra name to prepend to automatic file name.}

\item{...}{Extra arguments passed to plot creation function.}
}
\value{
Nothing.
}
\description{
Opens a new plot in the active session folder.  The plot will be named using the current time and the RLogBook params.
}
