% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{params}
\alias{params}
\title{Define an environment to hold package params and define some defaults}
\format{An object of class \code{environment} of length 10.}
\usage{
params
}
\description{
Define an environment to hold package params and define some defaults
}
\keyword{datasets}
