% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameSession.R
\name{renameSession}
\alias{renameSession}
\title{Rename an open session}
\usage{
renameSession(sessionName)
}
\arguments{
\item{sessionName}{The new session name to use.}
}
\description{
Renames an existing session.  This basically just adds a prefix to the existing session folder and updates all the sym-links.  It doesn't change the time stamp for the directory or anything else.
}
